+++ src/Application/Command/Happening/UpdateHandler.php
@@ -56,6 +56,7 @@ class UpdateHandler
             $update->limitParticipant,
             $update->isWebinar(),
             $update->isInteractiveWebinar(),
+            $update->isVideoWebinar(),
             $update->invitationCode,
             $update->liveUrl,
             $update->isWebinar() && $update->webinarRecorded