+++ src/Domain/Model/Happening.php
@@ -64,6 +64,9 @@ class Happening implements TimeRangeInterface
     /** @var bool */
     private $interactiveWebinar;
 
+    /** @var bool */
+    private $videoWebinar;
+
     /** @var null|string */
     private $webinarSessionId;
 
@@ -84,6 +87,7 @@ class Happening implements TimeRangeInterface
         ?string $invitationCode = null,
         bool $webinar = false,
         bool $interactiveWebinar = false,
+        bool $videoWebinar = false,
         ?string $liveUrl = null,
         bool $webinarRecorded = true
     ) {
@@ -102,6 +106,7 @@ class Happening implements TimeRangeInterface
         $this->products = new ArrayCollection();
         $this->webinar = $webinar;
         $this->interactiveWebinar = $interactiveWebinar;
+        $this->videoWebinar = $videoWebinar;
         $this->liveUrl = $liveUrl;
         $this->webinarRecorded = $webinarRecorded;
     }
@@ -198,6 +203,7 @@ class Happening implements TimeRangeInterface
         ?int $limitParticipant,
         bool $webinar,
         bool $interactiveWebinar,
+        bool $videoWebinar,
         ?string $invitationCode = null,
         ?string $liveUrl = null,
         bool $webinarRecorded = true
@@ -212,6 +218,7 @@ class Happening implements TimeRangeInterface
         $this->webinar = $webinar;
         $this->webinarRecorded = $webinarRecorded;
         $this->interactiveWebinar = $interactiveWebinar;
+        $this->videoWebinar = $videoWebinar;
         $this->liveUrl = $liveUrl;
     }
 
@@ -394,7 +401,7 @@ class Happening implements TimeRangeInterface
 
     public function isWebinar(): bool
     {
-        return $this->webinar || $this->interactiveWebinar;
+        return $this->webinar || $this->interactiveWebinar || $this->videoWebinar;
     }
 
     public function isInteractiveWebinar(): bool
@@ -402,8 +409,18 @@ class Happening implements TimeRangeInterface
         return $this->interactiveWebinar;
     }
 
+    public function isVideoWebinar(): bool
+    {
+        return $this->videoWebinar;
+    }
+
     public function getLiveUrl(): ?string
     {
         return $this->liveUrl;
+    }
+
+    public function isVideoWebinarAndHasLiveUrl(): bool
+    {
+        return $this->isVideoWebinar() && null !== $this->getLiveUrl();
     }
 }