+++ src/Infrastructure/Repository/Happening/Webinar/RecordArchiveRepository.php
@@ -3,6 +3,7 @@
 namespace Proximum\Vimeet\Infrastructure\Repository\Happening\Webinar;
 
 use Doctrine\ORM\EntityManager;
+use Proximum\Vimeet\Domain\Happening\Webinar\RecordStatus;
 use Proximum\Vimeet\Domain\Model\Happening;
 use Proximum\Vimeet\Domain\Model\Happening\Webinar\RecordArchive;
 use Proximum\Vimeet\Domain\Repository\Happening\Webinar\RecordArchiveRepositoryInterface;
@@ -38,6 +39,36 @@ class RecordArchiveRepository implements RecordArchiveRepositoryInterface
             ->setParameter('happening', $happening)
             ->getQuery()
             ->getResult()
+        ;
+    }
+
+    public function getByArchiveId(string $archiveId): ?RecordArchive
+    {
+        return $this->entityManager
+            ->createQueryBuilder()
+            ->select('recordArchive')
+            ->from(RecordArchive::class, 'recordArchive')
+            ->where('recordArchive.archiveId = :archiveId')
+            ->setParameter('archiveId', $archiveId)
+            ->setMaxResults(1)
+            ->getQuery()
+            ->getOneOrNullResult()
+        ;
+    }
+
+    public function hasStartedRecordArchiveForHappening(Happening $happening): bool
+    {
+        return null !== $this->entityManager
+            ->createQueryBuilder()
+            ->select('recordArchive.id')
+            ->from(RecordArchive::class, 'recordArchive')
+            ->where('recordArchive.happening = :happening')
+            ->andWhere('recordArchive.status = :status')
+            ->setParameter('happening', $happening)
+            ->setParameter('status', RecordStatus::STARTED)
+            ->setMaxResults(1)
+            ->getQuery()
+            ->getOneOrNullResult()
         ;
     }
 }